requirements.txt
gdown
langchain
langchain-community
sentence_transformers
chromadb
gpt4all
llama-cpp-python
langchainhub
accelerate
ragatouille